server/microservices/in_game_module/requirements.txt
functions-framework==3.*
flask==2.0.1
python-dotenv==1.0.0
flask_cors==3.0.10
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
boto3==1.28.19